[refactor] Change name boundary and reflected to out and in
Since when we consider the case where we would like to use pipe,
it does not make so much sense to use the name boundary and reflected
boundary, I just changed the name to out-boundary and in-boundary.
In the case of wall-collisions, the direction which bumps into the wall
will be the out-boundary.